convert to float32

+++ dianna/utils/onnx_runner.py
@@ -1,4 +1,5 @@
 import onnxruntime as ort
+import numpy as np
 
 
 class SimpleModelRunner:
@@ -29,6 +30,6 @@ class SimpleModelRunner:
         if self.preprocess_function is not None:
             input_data = self.preprocess_function(input_data)
 
-        onnx_input = {input_name: input_data}
+        onnx_input = {input_name: input_data.astype(np.float32)}
         pred_onnx = sess.run([output_name], onnx_input)[0]
         return pred_onnx
